==> 01-DataPreprocessing.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

googleDrivePathPrefix = 'drive/My Drive/Colab Notebooks'

from os import path
import re
import pandas as pd
import numpy as np


data_examples = []
with open(path.join(googleDrivePathPrefix,'data/cmn.txt'), 'r',encoding='utf8') as f:
  for line in f.readlines():
    data_examples.append(line)

english=[]
chinese=[]
for data in data_examples:
  splits=re.split('\t',data)
  english.append(splits[0])
  chinese.append(splits[1])

df = pd.DataFrame(data={'english':english,'chinese':chinese})

df.head()

df.count()

Filter texts with quotation marks `"`

- The intention is to remove cases where there is conversational nature in the the sentence, e.g.

df.loc[24690]

conversational_sample_mask = df['english'].str.contains('" "')
print(f'Conversational example counts: {conversational_sample_mask.sum()}')

df_conversational_samples = df[conversational_sample_mask]
df_conversational_samples.head()

df_filtered_conversation = df[~conversational_sample_mask]
print(f'Data sample counts (filtered conversations):\n{df_filtered_conversation.count()}')

df_filtered_conversation.sample(5)

Filter chinese text with english characters, numbers and non-essential punctuation marks excluding `[。，？！]`.

- To simplify separating individual characters in chinese texts, we filter out samples with english characters, numbers, and non-essential punctuation marks.
- Thus, we can simply split each character in the chinese texts and treat them as individual chinese character or punctuation mark.

chinese_with_eng_char_sample_mask = df['chinese'].str.contains("[A-Za-z0-9\-/•（）《》「」“”\[\]\"]")
chinese_with_eng_char_sample_mask.sum()

df_chinese_with_eng_char_sample = df_filtered_conversation[chinese_with_eng_char_sample_mask]
df_chinese_with_eng_char_sample.sample(5)

df_filtered_final = df_filtered_conversation[~chinese_with_eng_char_sample_mask]
print(f'Data sample counts (filtered chinese text with english characters):\n{df_filtered_final.count()}')

df_filtered_final.sample(5)

Replace chinese text with ascii punctuation marks, e.g. `[.,?!]` with chinese unicode punctuation marks `[。，？！]`.

chinese_with_ascii_punctuation_mask = df_filtered_final['chinese'].str.contains("[?.!,]")
chinese_with_ascii_punctuation_mask.sum()

df_chinese_with_ascii_punct = df_filtered_final[chinese_with_ascii_punctuation_mask]
ind = np.random.choice(df_chinese_with_ascii_punct.index,5)
df_chinese_with_ascii_punct.loc[ind[:]]

def replace_punct(sentence):
  sentence = re.sub(r"[?]", r"？", sentence)
  sentence = re.sub(r"[.]", r"。", sentence)
  sentence = re.sub(r"[!]", r"！", sentence)
  sentence = re.sub(r"[,]", r"，", sentence)
  return sentence

df_replace_punctuation = df_filtered_final
df_replace_punctuation['chinese'] = df_replace_punctuation['chinese'].apply(lambda x: replace_punct(x))
df_replace_punctuation[chinese_with_ascii_punctuation_mask].loc[ind[:]]

df_replace_punctuation['chinese'].str.contains("[?.!,]").sum()

Separate punctuation marks with additional space for the purpose of string splitting.

df_separate_punctuation = df_replace_punctuation.applymap(lambda x: re.sub(r"([.,?!。，？！])", r" \1 ", x))

df_separate_punctuation.sample(5)

Mapping function to do the text spliting.

- Split english text with space.
- Split chinese text at every character, discard the first and last element.

df_individual_char = df_separate_punctuation
df_individual_char['english_split']=df_individual_char['english'].map(lambda x: re.split(' ',x))
df_individual_char['chinese_split']=df_individual_char['chinese'].map(lambda x: re.split('',x)[1:-2])

#ind = np.random.choice(df_individual_char['english'].count(),5)
ind = np.array([23215,  7732, 22636,  6434, 20126])  # fix 
ind

df_individual_char.loc[ind]

Remove undesired characters in the text arrays.

- Remove '' in english text.
- Remove ' ' in chinese text.

def remove_char(x, char):
  x = np.asarray(x)
  mask = x == np.repeat(char, len(x))
  return x[~mask]


tmp=df_individual_char['english_split'].loc[ind[0]]
print(f'Raw input: {tmp}\n')
result = remove_char(tmp, '')
print(f'Removed char \'\': {result}')

tmp=df_individual_char['chinese_split'].loc[ind[2]]
print(f'Raw input: {tmp}\n')
result = remove_char(tmp, ' ')
print(f'Removed char \' \': {result}')

df_remove_array_char = df_individual_char
df_remove_array_char['english_split']=df_remove_array_char['english_split'].map(lambda x: remove_char(x,''))
df_remove_array_char['chinese_split']=df_remove_array_char['chinese_split'].map(lambda x: remove_char(x,' '))

df_remove_array_char.loc[ind]

Append `<start>` and `<end>` token to the chinese text.

def append_token(x):
  x=np.array(x,dtype='U7')
  start_token_added = np.insert(x,0,"<start>",axis=0)
  end_token_added = np.concatenate((start_token_added,["<end>"]),axis=0)
  return end_token_added

append_token(df_remove_array_char['chinese_split'].loc[ind[0]])

df_final = df_remove_array_char
df_final['chinese_split']=df_final['chinese_split'].map(lambda x: append_token(x))

for i in ind:
  print(f'{i}-th sample:\n')
  eng = df_final['english_split'].loc[i]
  chn = df_final['chinese_split'].loc[i]
  print(f'english: {eng}\n')
  print(f'chinese: {chn}\n\n')

Save the processed dataset.

df_final.to_json(path.join(googleDrivePathPrefix,'data/cmn-processed.json'))

df_reproduced = pd.read_json(path.join(googleDrivePathPrefix,'data/cmn-processed.json'))
df_reproduced.sample(5)
