==> 02-Tokenization.ipynb <==
%tensorflow_version 2.x
import tensorflow as tf
print(tf.__version__)

from google.colab import drive
drive.mount('/content/drive')

googleDrivePathPrefix = 'drive/My Drive/Colab Notebooks'

import pandas as pd
import numpy as np
from os import path
import json

from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

df = pd.read_json(path.join(googleDrivePathPrefix,'data/cmn-processed.json'))

df.head()

english = df['english_split']
chinese = df['chinese_split']
print(f'english-chinese sample count: {english.count()}')

tokenizer = Tokenizer(num_words=None,filters='',lower=False,char_level=False)

tokenizer_config = tokenizer.get_config()
tokenizer_config

tokenizer.fit_on_texts(chinese)
tokenizer_seq = tokenizer.texts_to_sequences(chinese)

tokenizer_config = tokenizer.get_config()

word_count_dict = json.loads(tokenizer_config['word_counts'])
df_word_count = pd.DataFrame({'char':word_count_dict.keys(), 'count':word_count_dict.values()})
print('Word count for top 10 highest frequency characters:')
df_word_count.sort_values(by=['count'],ascending=False).head(10)

print('Word count for 10 random samples:')
df_word_count.sample(10)

df_chinese_tokenized = df
df_chinese_tokenized['chinese_tokenized'] = tokenizer_seq

df_chinese_tokenized.head()

Save the tokenized dataset.

df_chinese_tokenized.to_json(path.join(googleDrivePathPrefix,'data/cmn-processed-tokenized.json'))

df_reproduced = pd.read_json(path.join(googleDrivePathPrefix,'data/cmn-processed-tokenized.json'))
df_reproduced.head()

Save tokenizer.

data = tokenizer.to_json()
with open(path.join(googleDrivePathPrefix,'data/tokenizer.json'), 'w') as f:
          json.dump(data, f)
